requirements.txt
flask-socketio
Flask-PyMongo
ipython